+++ lib/index.ts
@@ -366,7 +366,7 @@ export default class Qsu {
 		return str.trim().replace(removeAllSpace ? /\s+/g : /\s{2,}/g, '');
 	}
 
-	static removeSpecialChar(str: string, withoutSpace?: boolean): string {
+	static removeSpecialChar(str: string, exceptionCharacters?: string): string {
 		if (!str) {
 			return '';
 		}
@@ -374,7 +374,7 @@ export default class Qsu {
 		return str.replace(
 			new RegExp(
 				`[^a-zA-Z가-힣ㄱ-ㅎㅏ-ㅣ0-9\u3040-\u30ff\u3400-\u4dbf\u4e00-\u9fff\uf900-\ufaff\uff66-\uff9f${
-					withoutSpace ? ' ' : ''
+					exceptionCharacters ?? ''
 				}]`,
 				'gi'
 			),