+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.1.7 (2023-03-)
 
 - NodeJS 12 version deprecation
+- `removeSpecialChar`: Using `exceptionCharacters` instead of `withoutSpace`
 
 ## 1.1.6 (2023-02-28)